refactor (LottoTickets.java) : 강제 형변환 제거

+++ src/main/java/enums/Rank.java
@@ -27,7 +27,7 @@ public enum Rank {
         return matchCount;
     }
 
-    public Money getWinningPrize(int ticketCount) {
+    public Money getWinningPrize(long ticketCount) {
         return new Money(prize * ticketCount);
     }
